+++ packages/babel-jest/src/__tests__/index.js
@@ -4,7 +4,8 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
-const babelJest = require('../index');
+
+import babelJest from '../index';
 
 //Mock data for all the tests
 const sourceString = `