+++ packages/jest-mock/src/index.js
@@ -898,4 +898,4 @@ class ModuleMockerClass {
 }
 
 export type ModuleMocker = ModuleMockerClass;
-module.exports = new ModuleMockerClass(global);
+export default new ModuleMockerClass(global);