+++ packages/jest-snapshot/src/__tests__/utils.test.js
@@ -5,23 +5,22 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-jest.mock('fs');
-
-const fs = require('fs');
-const path = require('path');
-const chalk = require('chalk');
-
-const {
+import fs from 'fs';
+import path from 'path';
+import chalk from 'chalk';
+import {
+  deepMerge,
   getSnapshotData,
   keyToTestName,
   saveSnapshotFile,
   serialize,
-  testNameToKey,
-  deepMerge,
   SNAPSHOT_GUIDE_LINK,
   SNAPSHOT_VERSION,
   SNAPSHOT_VERSION_WARNING,
-} = require('../utils');
+  testNameToKey,
+} from '../utils';
+
+jest.mock('fs');
 
 const writeFileSync = fs.writeFileSync;
 const readFileSync = fs.readFileSync;